Part 7 complete
When viewing the repository, the views folder must contain an edit.jsx file within the places folder.
When going to localhost:3000/places/edit, it renders a form that allows a user to update a place.
When going to localhost:3000/places/edit, the form has the place's previous values set as default values in all inputs.
When going successfully submitting the edit form, the user is redirected to the place's show page with the updated information displayed.
When submitting the edit form without updating a value, the previous value remains the same when the place's show page is rendered.